third_party/WebKit/LayoutTests/fast/css/hover-recalc-expected.txt

Check that mousemove does not recalc style synchronously.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.internals is defined.
PASS window.eventSender is defined.
PASS getComputedStyle(t1, null).backgroundColor is transparent
PASS getComputedStyle(t2, null).backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS getComputedStyle(t1, null).backgroundColor is green
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS getComputedStyle(t2, null).backgroundColor is green
PASS getComputedStyle(t2.firstChild, null).width is '150px'
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS getComputedStyle(t3, null).backgroundColor is green
PASS getComputedStyle(t3.firstChild, null).width is '150px'
PASS successfullyParsed is true

TEST COMPLETE
